refactor: 문구 리팩토링 및 입력 값 유효 검증 사항 추가
