+++ src/main/kotlin/racingcar/view/attribute/ResultViewAttribute.kt
@@ -1,0 +1,5 @@
+package racingcar.view.attribute
+
+import racingcar.domain.Cars
+
+data class ResultViewAttribute(val cars: Cars)
