--- src/main/kotlin/racingcar/view/dto/ResultDto.kt
@@ -1,5 +1,0 @@
-package racingcar.view.dto
-
-import racingcar.domain.Cars
-
-data class ResultDto(val cars: Cars)
