+++ src/test/kotlin/racingcar/view/attribute/InputAttributeTest.kt
@@ -1,0 +1,6 @@
+package racingcar.view.attribute
+
+import io.kotest.core.spec.style.FunSpec
+
+class InputAttributeTest : FunSpec({
+})
